refactor: split chaordic and impulse types
